+++ src/main/java/models/dto/PagatEPunetoreve/CreatePagatEPunetoreveDto.java
@@ -47,7 +47,7 @@ public class CreatePagatEPunetoreveDto {
         this.viti = viti;
     }
 
-    public BigDecimal getShumaPaga() {
+    public double getShumaPaga() {
         return shumaPaga;
     }
